Authentication for user API key
* Added function in core package to check user api key
* Added django view to check user key in header

LAM-107 #time 3h #verify
