+++ webapp/webapp/urls.py
@@ -15,7 +15,9 @@ Including another URLconf
 """
 from django.conf.urls import include, url
 from django.contrib import admin
+from backend import views
 
 urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
+    url(r'^authenticate/', views.authenticate),
 ]
